Move client tests to participation
